Duct Tape Fix for Login Problems
Logging in works, but is not saved in a session and will logout as soon as you refresh. There is also no authorization checks anywhere, and all authentication is clientside. This is obviously not ideal but I don't have time to work on that now so I will come back to it later.